Change all the properties to match server json
The set of properties on the Car object were random bad names that I
had come up with. Now that I'm using the upstream json data structure,
it's worth just using those names to be less confusing. This is a big
breaking change, but better to do this early than late.

This also provides access to model information so that there is a
friendly name associated with each car.